+++ internals/repo/repo.go
@@ -1,3 +1,6 @@
 package repo
 
-const ()
+const (
+	// CorePackagesURL = "https://packages.stewproject.sh/"
+	CoreURL = "https://raw.githubusercontent.com/stewproject/core/main/packages/"
+)
